make cache not invalidate on non cached install

+++ src/Boot.tsx
@@ -60,9 +60,11 @@ window.addEventListener("load", async () => {
         ) {
             await anura.settings.set("milestone", milestone);
             await anura.settings.set("instancemilestone", instancemilestone);
-            await new Filer.fs.Shell().promises.rm("/anura_files", {
-                recursive: true,
-            });
+            if (anura.settings.get("use-sw-cache")) {
+                await new Filer.fs.Shell().promises.rm("/anura_files", {
+                    recursive: true,
+                });
+            }
             console.log("invalidated cache");
             window.location.reload();
         }
